requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==3.2.18
marshmallow==3.19.0
packaging==23.0
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
sqlparse==0.4.3
